Adding to data classes and updating UIs

+++ app/src/main/java/com/restaurantbooker/data/UserRepository.java
@@ -7,10 +7,9 @@ import com.restaurantbooker.user.User;
 
 public interface UserRepository {
         LiveData<Long> signupUser(UserEntity user);
-
         LiveData<UserEntity> getUserByEmail(String email);
         LiveData<Long> insertUser(UserEntity user);
-        LiveData<Integer> updateUser(UserEntity user);
+        LiveData<Long> updateUser(UserEntity user);
         LiveData<Long> deleteUser(UserEntity user);
         LiveData<UserEntity> signupUser(String email, String password);
 }
